+++ CHANGELOG.markdown
@@ -1,3 +1,7 @@
+next
+----
+* Added `Cons` and `Snoc` instances for `Control.Applicative.ZipList`
+
 4.16 [2018.01.28]
 -----------------
 * The `Semigroup` instances for `Traversed` and `Sequenced` are now more
@@ -32,8 +36,6 @@
   with GHC 8.4).
 
 * Add `Semigroup` and `Monoid` instances for `Indexing`.
-
-* Added `Cons` and `Snoc` instances for `Control.Applicative.ZipList`
 
 4.15.4
 ----